pulled time display function into separate component as well
